Rename `sourcecred admin` to `serve`
Serve is a more descriptive name; it so happens the server allows using
the admin dashboard, but it also is good for e.g. locally previewing the
instance. This also matches the nomenclature for the yarn command in the
template instance.

Test plan: Manual testing, also `yarn flow` passes.
